Moving code back to EntryPoint and naming modules to resources
