+++ resources/ApprovePage.js
